+++ src/main/java/com/sparta/delivery_app/domain/admin/adminuser/service/AdminUserService.java
@@ -3,20 +3,16 @@ package com.sparta.delivery_app.domain.admin.adminuser.service;
 import com.sparta.delivery_app.common.security.AuthenticationUser;
 import com.sparta.delivery_app.domain.admin.adminuser.dto.AdminUserResponseDto;
 import com.sparta.delivery_app.domain.admin.adminuser.dto.PageAdminUserResponseDto;
-import com.sparta.delivery_app.domain.commen.page.util.PageUtil;
 import com.sparta.delivery_app.domain.user.adapter.UserAdapter;
 import com.sparta.delivery_app.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static com.sparta.delivery_app.domain.commen.page.util.PageUtil.*;
+import static com.sparta.delivery_app.domain.common.Page.PageConstants.PAGE_SIZE_FIVE;
+import static com.sparta.delivery_app.domain.common.Page.PageUtil.*;
 import static com.sparta.delivery_app.domain.user.entity.UserStatus.checkManagerEnable;
 
 @Slf4j
