+++ src/main/java/com/sparta/delivery_app/domain/common/BaseTimeEntity.java
@@ -1,10 +1,11 @@
-package com.sparta.delivery_app.domain.commen;
+package com.sparta.delivery_app.domain.common;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
@@ -12,10 +13,13 @@ import java.time.LocalDateTime;
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseTimeCreateEntity {
+public abstract class BaseTimeEntity {
 
     @Column(updatable = false)
     @CreatedDate
     private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedAt;
 
 }
