+++ src/main/java/com/sparta/delivery_app/domain/common/Page/PageConstants.java
@@ -1,0 +1,13 @@
+package com.sparta.delivery_app.domain.common.Page;
+
+public class PageConstants {
+    // 페이지 크기
+    public static final Integer PAGE_SIZE_FIVE = 5;
+    public static final Integer PAGE_SIZE_TEN = 10;
+
+    // 정렬 기준
+    public static final String SORT_BY_CREATED_AT = "createdAt";
+
+    public static final String NO_ELEMENT_MESSAGE = "조회된 데이터가 없습니다.";
+
+}
