+++ src/main/java/com/sparta/delivery_app/domain/common/Page/PageUtil.java
@@ -1,4 +1,4 @@
-package com.sparta.delivery_app.domain.commen.page.util;
+package com.sparta.delivery_app.domain.common.Page;
 
 import com.sparta.delivery_app.common.exception.errorcode.PageErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.PageNotFoundException;
@@ -7,10 +7,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
+import static com.sparta.delivery_app.domain.common.Page.PageConstants.*;
+
 public class PageUtil {
-    public static final Integer PAGE_SIZE_FIVE = 5;
-    public static final Integer PAGE_SIZE_TEN = 10;
-    public static final String NO_ELEMENT_MESSAGE = "조회된 데이터가 없습니다.";
 
     public static Pageable createPageable(Integer pageNum, Integer pageSize, Boolean isDesc) {
         if (pageNum < 1) {
@@ -18,7 +17,7 @@ public class PageUtil {
         }
 
         Sort.Direction direction = isDesc ? Sort.Direction.DESC : Sort.Direction.ASC;
-        Sort sort = Sort.by(direction, "createdAt");
+        Sort sort = Sort.by(direction, SORT_BY_CREATED_AT);
 
         return PageRequest.of(pageNum - 1, pageSize, sort);
     }
