+++ src/main/java/com/sparta/delivery_app/domain/menu/entity/Menu.java
@@ -2,7 +2,7 @@ package com.sparta.delivery_app.domain.menu.entity;
 
 import com.sparta.delivery_app.common.exception.errorcode.OrderErrorCode;
 import com.sparta.delivery_app.common.globalcustomexception.StoreMenuMismatchException;
-import com.sparta.delivery_app.domain.commen.BaseTimeEntity;
+import com.sparta.delivery_app.domain.common.BaseTimeEntity;
 import com.sparta.delivery_app.domain.menu.dto.request.MenuAddRequestDto;
 import com.sparta.delivery_app.domain.menu.dto.request.MenuModifyRequestDto;
 import com.sparta.delivery_app.domain.store.entity.Store;
