+++ src/main/java/com/sparta/delivery_app/domain/review/entity/UserReviews.java
@@ -1,6 +1,6 @@
 package com.sparta.delivery_app.domain.review.entity;
 
-import com.sparta.delivery_app.domain.commen.BaseTimeEntity;
+import com.sparta.delivery_app.domain.common.BaseTimeEntity;
 import com.sparta.delivery_app.domain.liked.entity.ReviewLiked;
 import com.sparta.delivery_app.domain.order.entity.Order;
 import com.sparta.delivery_app.domain.review.dto.request.UserReviewModifyRequestDto;
